refactored testing suite and db_helpers to modularize the organization of the project. Continuing to unit test db/
