+++ backend/tests/conftest.py
@@ -11,10 +11,48 @@ logging.basicConfig(level=logging.DEBUG,
                     handlers=[logging.StreamHandler()])  # Log only to console
 logger = logging.getLogger("test_logger")
 
+@pytest.fixture(scope="function")
+def init_db():
+    """Fixture to setup a clean test database for each test."""
+    app = create_app()
+
+    logger.debug("Creating app and setting up the database...")
+    
+    with app.app_context():
+        # Verify if app context is properly set
+        logger.debug(f"App context set: {app.app_context() is not None}")
+        
+        # Check if db is correctly initialized and if the tables exist
+        logger.debug("Attempting to create all tables in the database.")
+        try:
+            db.create_all()  # Ensure tables are created within the app context
+            logger.debug("Database initialized for testing.")
+            print("Database initialized for testing.")
+        except Exception as e:
+            logger.error(f"Error during database setup: {e}")
+            print(f"Error during database setup: {e}")
+            raise  # Reraise exception to ensure the test fails
+
+    yield
+    
+    # Cleanup after the test
+    with app.app_context():
+        try:
+            logger.debug("Cleaning up the database...")
+            print("Cleaning up the database...")
+            db.session.remove()
+            db.drop_all()  # Drop all tables after test
+            logger.debug("Database cleaned up after test.")
+            print("Database cleaned up after test.")
+        except Exception as e:
+            logger.error(f"Error during database cleanup: {e}")
+            print(f"Error during database cleanup: {e}")
+            raise  # Reraise exception to ensure the test fails
+
 @pytest.fixture(scope="session", autouse=True)
 def setup_and_teardown_db():
-    """Fixture to set up and tear down the database."""
-    print("Creating the app and setting up the database...")
+    """Fixture to setup and tear down the database once for the whole test session."""
+    print("Setting up the database...")
     app = create_app()
 
     logger.debug("App created. Proceeding with database initialization.")
